+++ jscomp/core/lam.ml
@@ -748,19 +748,6 @@ let sequor l r = if_ l true_ r
 (** [l && r ] *)
 let sequand l r = if_ l r false_
 
-(******************************************************************)
-(* only [handle_bs_non_obj_ffi] will be used outside *)
-(*
-   [no_auto_uncurried_arg_types xs]
-   check if the FFI have @uncurry attribute.
-   if it does not we wrap it in a nomral way otherwise
-*)
-let rec no_auto_uncurried_arg_types (xs : External_arg_spec.params) =
-  match xs with
-  | [] -> true
-  | { arg_type = Fn_uncurry_arity _ } :: _ -> false
-  | _ :: xs -> no_auto_uncurried_arg_types xs
-
 let result_wrap loc (result_type : External_ffi_types.return_wrapper) result =
   match result_type with
   | Return_replaced_with_unit -> seq result unit
@@ -771,31 +758,7 @@ let result_wrap loc (result_type : External_ffi_types.return_wrapper) result =
       prim ~primitive:Pundefined_to_opt ~args:[ result ] loc
   | Return_unset | Return_identity -> result
 
-let rec transform_uncurried_arg_type loc (arg_types : External_arg_spec.params)
-    (args : t list) =
-  match (arg_types, args) with
-  | { arg_type = Fn_uncurry_arity n; arg_label } :: xs, y :: ys ->
-      let o_arg_types, o_args = transform_uncurried_arg_type loc xs ys in
-      ( { External_arg_spec.arg_type = Nothing; arg_label } :: o_arg_types,
-        prim ~primitive:(Pjs_fn_make n) ~args:[ y ] loc :: o_args )
-  | x :: xs, y :: ys -> (
-      match x with
-      | { arg_type = Arg_cst _ } ->
-          let o_arg_types, o_args = transform_uncurried_arg_type loc xs args in
-          (x :: o_arg_types, o_args)
-      | _ ->
-          let o_arg_types, o_args = transform_uncurried_arg_type loc xs ys in
-          (x :: o_arg_types, y :: o_args))
-  | ([], [] | _ :: _, [] | [], _ :: _) as ok -> ok
-
 let handle_bs_non_obj_ffi (arg_types : External_arg_spec.params)
     (result_type : External_ffi_types.return_wrapper) ffi args loc prim_name ~dynamic_import =
-  if no_auto_uncurried_arg_types arg_types then
     result_wrap loc result_type
       (prim ~primitive:(Pjs_call { prim_name; arg_types; ffi; dynamic_import }) ~args loc)
-  else
-    let n_arg_types, n_args = transform_uncurried_arg_type loc arg_types args in
-    result_wrap loc result_type
-      (prim
-         ~primitive:(Pjs_call { prim_name; arg_types = n_arg_types; ffi; dynamic_import })
-         ~args:n_args loc)