+++ jscomp/frontend/external_arg_spec.ml
@@ -54,7 +54,6 @@ type attr =
   (* `a does not have any value*)
   | Int of (string * int) list (* ([`a | `b ] [@int])*)
   | Arg_cst of cst
-  | Fn_uncurry_arity of int (* annotated with [@uncurry ] or [@uncurry 2]*)
   (* maybe we can improve it as a combination of {!Asttypes.constant} and tuple *)
   | Extern_unit
   | Nothing